Include cmath if std::nan is used
In the Localization2DServer.cpp file, the std::nan function is used. This function is defined in the cmath header ( https://en.cppreference.com/w/cpp/numeric/math/nan ).
Before this commit the math.h header (that only defines nan in the global namespace, not std::nan) was included instead of cmath.
This can create a compilation failure or not depending on the platform due to transitive inclusion of cmath in the other included headers.